backend/requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
WeasyPrint==60.1
email-validator==2.1.0
python-dotenv==1.0.0